arm: dts: overlays: pitft35-resistive: add upstream compatible
commit  5114fc55fc0aab729bcdf9567d96f7b176a41806 from
https://github.com/raspberrypi/linux.git rpi-5.4.y

The upstream hx8357d driver uses "adafruit,yx350hv15" for the compatible
string explicitly for this screen config and not a hx8357d generic for
the controller so add that in as well so it will work with an unmodified
upstream kernel driver. We leave the downstream as the priority.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>