drm/modes: parse_cmdline: Stop parsing extras after bpp / refresh at ', '
commit  902e887cf1d329a28b339aa4ec6214dd6d1cbff2 from
https://github.com/raspberrypi/linux.git rpi-5.4.y

Commit c2ed3e941901810ad3d55ce1935fa22c5007fee4 upstream.

Before this commit it was impossible to add an extra mode argument after
a bpp or refresh specifier, combined with an option, e.g.
video=HDMI-1:720x480-24e,rotate=180 would not work, either the "e" to
force enable would need to be dropped or the ",rotate=180", otherwise
the mode_option would not be accepted.

This commit fixes this by fixing the length calculation if extras_ptr
is set to stop the extra parsing at the start of the options (stop at the
',' options_ptr points to).

Acked-by: Maxime Ripard <mripard@kernel.org>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20191118155134.30468-3-hdegoede@redhat.com
Signed-off-by: Meng Li <Meng.Li@windriver.com>