MLK-21599-1 arm64: Kconfig: Make FORCE_MAX_ZONEORDER configurable
commit  from
https://source.codeaurora.org/external/imx/linux-imx

Older imx releases increased this based on SOC_IMX8 but that symbol is
gone and the expectation is that SOC-selection symbols like "ARCH_MXC"
will almost always be defined and shouldn't be used to make incompatible
config decisions.

Make the value of FORCE_MAX_ZONEORDER configurable so this can be
adjusted in a .config file.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Reviewed-by: Li Yang <Leoyang.li@nxp.com>
Signed-off-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>