LF-397-1 clk: imx8qxp: fix orphan clock rate of sdhc0_clk
commit  from
https://source.codeaurora.org/external/imx/linux-imx

Currently sdhc0_clk is an orphan clock and it's rate is wrong.
The reason is missing parent pll clocks defined in DT.
$ cat /sys/kernel/debug/clk/clk_summary  | grep sdhc
 ...
 sdhc0_clk                            0        0        0           0          0     0  50000
    sdhc0_lpcg_per_clk                0        0        0   396000000          0     0  50000

As sdhc does not have requiremnt to change the clock parent to AVPLL now,
let's remove the unused parent clock definition in clock driver to avoid
this issue.

Reviewed-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>